[numerics] simplify test ConvexQP
