+++ numerics/numerics_tests.cmake
@@ -420,72 +420,6 @@ if(WITH_${COMPONENT}_TESTING)
 
   NEW_GMP_TEST_COLLECTION(TEST_NSGS_COLLECTION_1)
 
-  
-  # # Warning. Only test GMP4 GMP5 have fc3d problem inside
-
-  # NEW_GMP_TEST(SICONOS_FRICTION_3D_ONECONTACT_QUARTIC GMP0.dat)
-  # NEW_GMP_TEST(SICONOS_FRICTION_3D_ONECONTACT_QUARTIC GMP1.dat)
-  # NEW_GMP_TEST(SICONOS_FRICTION_3D_ONECONTACT_QUARTIC GMP2.dat)
-  # NEW_GMP_TEST(SICONOS_FRICTION_3D_ONECONTACT_QUARTIC GMP3.dat)
-  # NEW_GMP_TEST(SICONOS_FRICTION_3D_ONECONTACT_QUARTIC GMP4.dat)
-  # NEW_GMP_TEST(SICONOS_FRICTION_3D_ONECONTACT_QUARTIC GMP5.dat)
-  # SET(test-GMP-REDUCED0_3D_ONECONTACT_QUARTIC-GMP6_PROPERTIES WILL_FAIL TRUE)
-  # NEW_GMP_TEST(SICONOS_FRICTION_3D_ONECONTACT_QUARTIC GMP6.dat)
-
-  # NEW_GMP_TEST(SICONOS_FRICTION_3D_ONECONTACT_NSN_GP GMP0.dat)
-  # NEW_GMP_TEST(SICONOS_FRICTION_3D_ONECONTACT_NSN_GP GMP1.dat)
-  # NEW_GMP_TEST(SICONOS_FRICTION_3D_ONECONTACT_NSN_GP GMP2.dat)
-  # NEW_GMP_TEST(SICONOS_FRICTION_3D_ONECONTACT_NSN_GP GMP3.dat)
-  # NEW_GMP_TEST(SICONOS_FRICTION_3D_ONECONTACT_NSN_GP GMP4.dat)
-  # NEW_GMP_TEST(SICONOS_FRICTION_3D_ONECONTACT_NSN_GP GMP5.dat)
-
-  # NEW_GMP_TEST(SICONOS_FRICTION_3D_ONECONTACT_NSN GMP4.dat)
-  # NEW_GMP_TEST(SICONOS_FRICTION_3D_ONECONTACT_NSN GMP5.dat)
-
-  # NEW_GMP_TEST(SICONOS_FRICTION_3D_ONECONTACT_ProjectionOnConeWithLocalIteration GMP4.dat)
-  # NEW_GMP_TEST(SICONOS_FRICTION_3D_ONECONTACT_ProjectionOnConeWithLocalIteration GMP5.dat)
-
-  # IF(HAS_LAPACK_DGESVD)
-  #   NEW_GMP_TEST(SICONOS_FRICTION_3D_ONECONTACT_QUARTIC GMP0.dat 0 0 0 0 0 1)
-  #   NEW_GMP_TEST(SICONOS_FRICTION_3D_ONECONTACT_QUARTIC GMP1.dat 0 0 0 0 0 1)
-  #   NEW_GMP_TEST(SICONOS_FRICTION_3D_ONECONTACT_QUARTIC GMP2.dat 0 0 0 0 0 1)
-  #   NEW_GMP_TEST(SICONOS_FRICTION_3D_ONECONTACT_QUARTIC GMP3.dat 0 0 0 0 0 1)
-  #   NEW_GMP_TEST(SICONOS_FRICTION_3D_ONECONTACT_QUARTIC GMP4.dat 0 0 0 0 0 1)
-  #   NEW_GMP_TEST(SICONOS_FRICTION_3D_ONECONTACT_QUARTIC GMP5.dat 0 0 0 0 0 1)
-  #   #    SET(test-GMP-REDUCED1_3D_QUARTIC-GMP6_PROPERTIES WILL_FAIL TRUE)
-  #   NEW_GMP_TEST(SICONOS_FRICTION_3D_ONECONTACT_QUARTIC GMP6.dat 0 0 0 0 0 1)
-  # endif()
-
-  # NEW_GMP_TEST(SICONOS_FRICTION_3D_ONECONTACT_QUARTIC GMP0.dat 0 0 0 0 0 2)
-  # NEW_GMP_TEST(SICONOS_FRICTION_3D_ONECONTACT_QUARTIC GMP1.dat 0 0 0 0 0 2)
-  # NEW_GMP_TEST(SICONOS_FRICTION_3D_ONECONTACT_QUARTIC GMP2.dat 0 0 0 0 0 2)
-  # NEW_GMP_TEST(SICONOS_FRICTION_3D_ONECONTACT_QUARTIC GMP3.dat 0 0 0 0 0 2)
-  # NEW_GMP_TEST(SICONOS_FRICTION_3D_ONECONTACT_QUARTIC GMP4.dat 0 0 0 0 0 2)
-  # NEW_GMP_TEST(SICONOS_FRICTION_3D_ONECONTACT_QUARTIC GMP5.dat 0 0 0 0 0 2)
-  # NEW_GMP_TEST(SICONOS_FRICTION_3D_ONECONTACT_QUARTIC GMP6.dat 0 0 0 0 0 2)
-
-  # NEW_GMP_TEST(SICONOS_FRICTION_3D_ONECONTACT_QUARTIC GMP0.dat 0 0 0 0 0 3)
-  # NEW_GMP_TEST(SICONOS_FRICTION_3D_ONECONTACT_QUARTIC GMP1.dat 0 0 0 0 0 3)
-  # NEW_GMP_TEST(SICONOS_FRICTION_3D_ONECONTACT_QUARTIC GMP2.dat 0 0 0 0 0 3)
-  # NEW_GMP_TEST(SICONOS_FRICTION_3D_ONECONTACT_QUARTIC GMP3.dat 0 0 0 0 0 3)
-  # NEW_GMP_TEST(SICONOS_FRICTION_3D_ONECONTACT_QUARTIC GMP6.dat 0 0 0 0 0 3)
-
-  # NEW_GMP_TEST(SICONOS_FRICTION_3D_ONECONTACT_NSN_GP GMP4.dat 0 0 0 0 0 2)
-  # NEW_GMP_TEST(SICONOS_FRICTION_3D_ONECONTACT_NSN_GP GMP5.dat 0 0 0 0 0 2)
-  # NEW_GMP_TEST(SICONOS_FRICTION_3D_ONECONTACT_NSN_GP GMP6.dat 0 0 0 0 0 2)
-
-  # NEW_GMP_TEST(SICONOS_FRICTION_3D_ONECONTACT_NSN GMP4.dat 0 0 0 0 0 2)
-  # NEW_GMP_TEST(SICONOS_FRICTION_3D_ONECONTACT_NSN GMP5.dat 0 0 0 0 0 2)
-  # NEW_GMP_TEST(SICONOS_FRICTION_3D_ONECONTACT_NSN GMP6.dat 0 0 0 0 0 2)
-
-  # NEW_GMP_TEST(SICONOS_FRICTION_3D_ONECONTACT_ProjectionOnConeWithLocalIteration GMP4.dat 0 0 0 0 0 2)
-  # NEW_GMP_TEST(SICONOS_FRICTION_3D_ONECONTACT_ProjectionOnConeWithLocalIteration GMP5.dat 0 0 0 0 0 2)
-  # NEW_GMP_TEST(SICONOS_FRICTION_3D_ONECONTACT_ProjectionOnConeWithLocalIteration GMP6.dat 0 0 0 0 0 2)
-
-
-
-
-
   END_TEST()
   #BEGIN_TEST(src/GenericMechanical/test)
   #NEW_TEST(GMP_FAILED GenericMechanical_test1.c)
@@ -522,15 +456,11 @@ if(WITH_${COMPONENT}_TESTING)
   ENDFOREACH()
   END_TEST()
   BEGIN_TEST(src/QP/test)
-  NEW_TEST(ConvexQP_test0 ConvexQP_test.c)
-  NEW_TEST(ConvexQP_PG ConvexQP_test1.c)
-  NEW_TEST(ConvexQP_ADDM ConvexQP_test3.c)
-  NEW_TEST(ConvexQP_ADDM_ACCELERATION ConvexQP_test4.c)
-  NEW_TEST(ConvexQP_ADDM_ACCELERATION_AND_RESTART ConvexQP_test5.c)
-  NEW_TEST(ConvexQP_PG_FC3D ConvexQP_testFC3D2.c)
-  NEW_TEST(ConvexQP_ADMM_FC3D ConvexQP_testFC3D3.c)
-  NEW_TEST(ConvexQP_ADMM_ACCELERATION_FC3D ConvexQP_testFC3D4.c)
-  NEW_TEST(ConvexQP_ADMM_ACCELERATION_AND_RESTART_FC3D ConvexQP_testFC3D5.c)
+
+  NEW_TEST(ConvexQP_test_collection ConvexQP_test.c)
+
+  NEW_TEST(ConvexQP_FC3D_test_collection ConvexQP_FC3D_test.c)
+ 
   END_TEST()
 
   # NEW_TEST(ConvexQP_ADMM_reduced ConvexQP_test2.c)
